Change test framework
