+++ .vscode/launch.json
@@ -30,6 +30,7 @@
 				"--extensionDevelopmentPath=${workspaceFolder}"
 			],
 			"env": {
+				"AZCODE_DOCKER_IGNORE_BUNDLE": "",
 				"DEBUGTELEMETRY": "1"
 			},
 			"stopOnEntry": false,
@@ -37,7 +38,7 @@
 			"outFiles": [
 				"${workspaceFolder}/out/**/*.js"
 			],
-			"preLaunchTask": "npm: webpack"
+			"preLaunchTask": "npm: webpack-prod"
 		},
 		{
 			"name": "Launch Tests",
@@ -47,7 +48,7 @@
 			"args": [
 				"${workspaceFolder}/test/test.code-workspace",
 				"--extensionDevelopmentPath=${workspaceFolder}",
-				"--extensionTestsPath=${workspaceFolder}/out/test"
+				"--extensionTestsPath=${workspaceFolder}/out/test/index"
 			],
 			"stopOnEntry": false,
 			"sourceMaps": true,
@@ -58,9 +59,6 @@
 			"env": {
 				"AZCODE_DOCKER_IGNORE_BUNDLE": "1",
 				"MOCHA_grep": "", // RegExp of tests to run (empty means all)
-				"MOCHA_enableTimeouts": "0", // Disable time-outs
-				"DEBUGTELEMETRY": "1",
-				"NODE_DEBUG": ""
 			}
 		},
 		{
@@ -71,7 +69,7 @@
 			"args": [
 				"${workspaceFolder}/test/test.code-workspace",
 				"--extensionDevelopmentPath=${workspaceFolder}",
-				"--extensionTestsPath=${workspaceFolder}/out/test"
+				"--extensionTestsPath=${workspaceFolder}/out/test/index"
 			],
 			"stopOnEntry": false,
 			"sourceMaps": true,
@@ -82,9 +80,6 @@
 			"env": {
 				"AZCODE_DOCKER_IGNORE_BUNDLE": "1",
 				"MOCHA_grep": "\\(unit\\)", // RegExp of tests to run (empty means all)
-				"MOCHA_enableTimeouts": "0", // Disable time-outs
-				"DEBUGTELEMETRY": "1",
-				"NODE_DEBUG": ""
 			}
 		},
 		{
@@ -95,19 +90,17 @@
 			"args": [
 				"${workspaceFolder}/test/test.code-workspace",
 				"--extensionDevelopmentPath=${workspaceFolder}",
-				"--extensionTestsPath=${workspaceFolder}/dist/test"
+				"--extensionTestsPath=${workspaceFolder}/dist/test/index"
 			],
 			"stopOnEntry": false,
 			"sourceMaps": true,
 			"outFiles": [
 				"${workspaceFolder}/dist/**/*.js"
 			],
-			"preLaunchTask": "npm: webpack",
+			//"preLaunchTask": "npm: webpack-prod",
 			"env": {
+				"AZCODE_DOCKER_IGNORE_BUNDLE": "",
 				"MOCHA_grep": "", // RegExp of tests to run (empty means all)
-				"MOCHA_enableTimeouts": "0", // Disable time-outs
-				"DEBUGTELEMETRY": "1",
-				"NODE_DEBUG": ""
 			}
 		}
 	]
